Change centroid calculation to do online update and return doubles

+++ filters/VoxelCentroidNearestNeighborFilter.cpp
@@ -94,7 +94,7 @@ PointViewSet VoxelCentroidNearestNeighborFilter::run(PointViewPtr view)
     PointViewPtr output = view->makeNew();
     for (auto const& t : populated_voxel_ids)
     {
-        Eigen::Vector3f centroid = eigen::computeCentroid(*view, t.second);
+        Eigen::Vector3d centroid = eigen::computeCentroid(*view, t.second);
         std::vector<PointId> neighbors =
             kdi.neighbors(centroid[0], centroid[1], centroid[2], 1);
         output->appendPoint(*view, neighbors[0]);
